% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus.R
\name{consensus}
\alias{consensus}
\title{Takes the most common nucleotide in each position of a group motifs}
\usage{
consensus(Motifs)
}
\arguments{
\item{Motifs}{A vector of character string of the DNA Motifs.}
}
\value{
A vector of the most common nucleotide in each position
}
\description{
Takes the most common nucleotide in each position of a group motifs
}
