% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DosR_Dataset}
\alias{DosR_Dataset}
\title{Dormacy Survival regulartors (DosR) upstream genes}
\format{
A vector of character strings of the 10 genes
}
\source{
\url{http://bioinformaticsalgorithms.com/data/challengedatasets/DosR.txt}
}
\usage{
DosR_Dataset
}
\description{
These are 10 genes that express significant
changes during hypoxic condition. For our example we can use these data to see if we can
determine possible Motifs that may control gene expression.
}
\keyword{datasets}
