% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{frequencyMap}
\alias{frequencyMap}
\title{FrequencyMap}
\usage{
frequencyMap(k, txt)
}
\arguments{
\item{k}{An Integer representing length of kmer}

\item{txt}{A character string of DNA sequence}
}
\value{
A List of frequency of kmers in txt sequence, Class of object is frequencyMap allowing
for specifc plot method.
}
\description{
Extract frequency of specific kmers (snippets of specific length) from a larger sequence.
}
