% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GreedyMotifSearch.R
\name{GreedyMotifSearch}
\alias{GreedyMotifSearch}
\title{Given a series of DNA sequences this function will return the most probably motif in each sequence taking into account
the other DNA sequences.This uses a 'Greedy" method that is very computationally simple allowing for quick fast results.}
\usage{
GreedyMotifSearch(DNAseqs, k, t, pseudo = FALSE)
}
\arguments{
\item{DNAseqs}{A vector of character strings representing a series of DNA sequences}

\item{k}{An integer Value representing how long of kmer ins interested}

\item{t}{An integer value representing how many of the DNAseqs one wish to incorporate}

\item{pseudo}{A TRUE or FALSE condition, default FALSE. If TRUE will implement pseudo counts which alleviates biases
against some single point mutations}
}
\value{
A vector of best scoring Motifs
}
\description{
Given a series of DNA sequences this function will return the most probably motif in each sequence taking into account
the other DNA sequences.This uses a 'Greedy" method that is very computationally simple allowing for quick fast results.
}
