% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muCount.R
\name{muCount}
\alias{muCount}
\title{A function that takes a series of motifs and return a count of each AGTC at each position.}
\usage{
muCount(Motifs, pseudo = FALSE)
}
\arguments{
\item{Motifs}{vector of motifs of interest}

\item{pseudo}{TRUE or FALSE statement, if pseudo is true than will implement pseudo counts.}
}
\value{
A matrix with labeled Nucleotide
}
\description{
A function that takes a series of motifs and return a count of each AGTC at each position.
}
\examples{
muCount(sampleMotifs,pseudo=TRUE)
}
