% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muProfile.R
\name{muProfile}
\alias{muProfile}
\title{muProfile}
\usage{
muProfile(Motifs, pseudo = FALSE)
}
\arguments{
\item{Motifs}{Vector of character DNA sequences}

\item{pseudo}{TRUE or FALSE condition, default FALSE. If TRUE will implement pseudo counts which alleviates biases
against some single point mutations}
}
\value{
A profile class object with
}
\description{
Generates a probability profile based on a series of sequences 'Motifs'
}
\examples{
muProfile(sampleMotif)
}
