% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muScore.R
\name{muScore}
\alias{muScore}
\title{muScore}
\usage{
muScore(Motifs)
}
\arguments{
\item{Motifs}{A vector of character string of the DNA Motifs.}
}
\value{
A numeric value representing a score of the consensus based on Motifs.
The lower the score the stronger the consensus compares to Motifs
}
\description{
This function gives a score of how the consensus nucleotide compares to the other in a motif.
}
\examples{
muScore(sampleMotifs)
}
