% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patternCount.R
\name{patternCount}
\alias{patternCount}
\title{Given a pattern will count how many times the pattern occurs in a line of text.}
\usage{
patternCount(patt, txt)
}
\arguments{
\item{patt}{A character string}

\item{txt}{A character string}
}
\value{
The number of times \code{patt} occurs in \code{txt}
}
\description{
Given a pattern will count how many times the pattern occurs in a line of text.
}
\examples{
patternCount("AGC","GTGCAGCTAGC")
}
