% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.frequencyMap.R
\name{plot.frequencyMap}
\alias{plot.frequencyMap}
\title{Method for generic plot function that visualize different kmer frequency}
\usage{
\method{plot}{frequencyMap}(frequencyMap)
}
\arguments{
\item{frequencyMap}{An object with the frequencyMap class (a List)}
}
\value{
a ggplot bargraph of the different kmers and their frequency
}
\description{
Method for generic plot function that visualize different kmer frequency
}
\examples{
plot(sampleFrequency)
}
