% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProfile.R
\name{plot.profile}
\alias{plot.profile}
\title{Method for generating sequence logo graphical representation for profile class objects.}
\usage{
\method{plot}{profile}(profile, prob = TRUE)
}
\arguments{
\item{profile}{A profile class Matrix.}

\item{prob}{A TRUE or FALSE argument. Default TRUE, will present visualization based on
probability. If FALSE will present based on weight of each sequence.}
}
\value{
A ggseqlogo generated sequence logo plot.
}
\description{
Method for generating sequence logo graphical representation for profile class objects.
}
\examples{
plot(sampleProfile)
}
\seealso{
[muMotif::muProfile()] creates profile class matrix
}
