% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr.R
\name{pr}
\alias{pr}
\title{pr}
\usage{
pr(txt, profile)
}
\arguments{
\item{txt}{A character string reprenting the DNA sequence in question.}

\item{profile}{A matrix representing the probability of each nucleuotide in specific locations}
}
\value{
A number value of the probability of the txt sequence based on the profile matrix
}
\description{
From a profile return the probability a DNA sequence fits from the profile
}
\examples{
pr("ACTG",sampleProfile)
}
