% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profileMostProbable.R
\name{profileMostProbable}
\alias{profileMostProbable}
\title{profileMostProbable}
\usage{
profileMostProbable(txt, k, profile)
}
\arguments{
\item{txt}{A character string of a DNA sequence that one wants to know most probable kmer}

\item{k}{A numeric value of the length of a kmer}

\item{profile}{A matrix of a profile that will determine most probable kmer}
}
\value{
A character string of the DNA kmer most probable
}
\description{
With a profile, calculate most probable kmer in a specific sequence(txt)
}
\examples{
profileMostProbable("GTCGTCA",3,sampleProfile)
}
