% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simpleSampleMotif}
\alias{simpleSampleMotif}
\title{Simple sample motif data}
\format{
A vector of 5 motifs 12 sequences long
}
\usage{
simpleSampleMotif
}
\description{
A collection of Motifs that are small and simple enough to showcase functions in muMotif
}
\keyword{datasets}
